# src/MyBudget.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MyBudget",
    "description": "",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "",
      "email": ""
    },
    "license": {
      "name": "Use under MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "v1"
  },
  "paths": {
    "/api/CostCenter/{costCenterId}": {
      "get": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Get CostCenter by id.",
        "description": "Get CostCenter by id.",
        "operationId": "getCostCenterById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "costCenterId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "schema": {
              "type": "string"
            }
          },
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCostCenterByIdResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Delete CostCenter.",
        "description": "Delete CostCenter.",
        "operationId": "removeCostCenter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "costCenterId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/RemoveCostCenterResponse"
            }
          }
        }
      }
    },
    "/api/CostCenter": {
      "get": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Get CostCenters.",
        "description": "Get CostCenters.",
        "operationId": "getCostCenters",
        "produces": [
          "application/json"
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCostCentersResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Create CostCenter.",
        "description": "Create CostCenter.",
        "operationId": "createCostCenter",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCostCenterRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CreateCostCenterResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Update CostCenter.",
        "description": "Update CostCenter.",
        "operationId": "updateCostCenter",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCostCenterRequest"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/UpdateCostCenterResponse"
            }
          }
        }
      }
    },
    "/api/CostCenter/page/{pageSize}/{index}": {
      "get": {
        "tags": [
          "CostCenter"
        ],
        "summary": "Get CostCenter Page.",
        "description": "Get CostCenter Page.",
        "operationId": "getCostCentersPage",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "pageSize",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "500": {
            "description": "Server Error"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GetCostCentersPageResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "CostCenterDto": {
      "type": "object",
      "properties": {
        "costCenterId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "CreateCostCenterRequest": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        }
      }
    },
    "CreateCostCenterResponse": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetCostCenterByIdResponse": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetCostCentersPageResponse": {
      "type": "object",
      "properties": {
        "length": {
          "format": "int32",
          "type": "integer"
        },
        "entities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CostCenterDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "GetCostCentersResponse": {
      "type": "object",
      "properties": {
        "costCenters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CostCenterDto"
          }
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "ProblemDetails": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "format": "int32",
          "type": "integer"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      },
      "additionalProperties": { }
    },
    "RemoveCostCenterResponse": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "UpdateCostCenterRequest": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        }
      }
    },
    "UpdateCostCenterResponse": {
      "type": "object",
      "properties": {
        "costCenter": {
          "$ref": "#/definitions/CostCenterDto"
        },
        "validationErrors": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}